+++ src/components/Rewards/editOffer.js
@@ -26,8 +26,8 @@ export default function EditReward({ businessID, currentReward, triggerUpdate, s
         let formObject = Object.fromEntries(data.entries());
         formObject.rewardCost = parseInt(formObject.rewardCost);
         formObject.rewardExpiryDuration = parseInt(formObject.rewardExpiryDuration);
-        formObject.rewardLat = parseInt(formObject.rewardLat);
-        formObject.rewardLon = parseInt(formObject.rewardLon);
+        formObject.rewardLat = parseFloat(formObject.rewardLat);
+        formObject.rewardLon = parseFloat(formObject.rewardLon);
         
         editReward({businessID: businessID, ...formObject}, onCompletion);
     }
